desktop: fix tap coordinates + add self-contained Windows build
Fix the input mapping: with a renderer logical size set, SDL already reports
mouse *event* coordinates in the logical (320x240) space, so the extra
SDL_RenderWindowToLogical call was scaling them a second time. Clicks landed at
1/scale of the cursor — feeding dropped food in the wrong place and the HUD/
Settings buttons couldn't be hit at any --scale > 1 (it only worked by accident
at scale 1, on every platform). Use the event coords directly; scale only the
normalized touch (FINGERDOWN) coords to the canvas.

Windows build is now one command (no vcpkg / manual SDL needed): when no system
SDL2 is found, CMake fetches and statically builds SDL2 2.32.4 via FetchContent,
so the result is a standalone .exe. Build it as a GUI (windowed) subsystem app
so launching it doesn't pop a console — important for kiosk use. Linux keeps
using the system libsdl2-dev.

Also declare per-monitor-v2 DPI awareness before SDL_Init so a Windows display
scaled above 100% renders crisply in real pixels instead of being virtualized
(harmless on other platforms), and fix two MSVC warnings (int->float cast;
the CRT strncpy/getenv nags via _CRT_SECURE_NO_WARNINGS).

Verified: native Windows MSVC build runs, window opens, feeding lands at the
cursor, and the HUD/Settings buttons respond; Linux build still clean.

+++ desktop/src/sim/Aquarium.cpp
@@ -923,7 +923,7 @@ void Aquarium::drawSeaweed(Framebuffer& fb) const {
         float bladeHeight = clampVal(32.0f * seaweedLength_ * heightVariation, 18.0f, 72.0f);
         int y0 = kScreenH - 2;
 
-        float prevX = bx, prevY = y0;
+        float prevX = bx, prevY = static_cast<float>(y0);
         const int segments = 7;
         for (int seg = 1; seg <= segments; ++seg) {
             float u = static_cast<float>(seg) / segments;
